requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
langchain
langchain-anthropic
langchain-core
python-multipart
boto3
cryptography
PyJWT
requests
